Distinguish IEEE 754 comparisons in IR
Currently IEEE 754 comparison is used only for binary operator
expressions as a part of operator convention resolution for ==, !=,
<, <=, >, >=:
IEEE 754 comparison is used, if both operands are statically known
to be of the same floating-point type (Float or Double).
Otherwise, operands are compared using 'equals' / 'compareTo' methods
(which in case of floating-point numbers implement a total ordering,
which distinguishes +-0.0 and NaNs).

See https://jetbrains.quip.com/4FKoATYO06uF

Simple way to check that is to compare '0.0' and '-0.0':
according to IEEE 754 rules:
  -0.0 == 0.0   is true,
  -0.0 < 0.0    is false
and according to total order implemented by java.lang.{Float, Double}:
  -0.0 == 0.0   is false,
  -0.0 < 0.0    is true

"Operand is statically known to be of type T" means that either the
expression has type T, or a smart cast to T is possible.
Note that technically there may be an expression that is "statically
known to be both Float and Double", e.g., 'x' in the last clause of
  x is Float && x is Double && x == 0.0
However, there's no actual value that fulfills such a condition, because
Float and Double are final classes and can't have common children.